Equipment Requirements
Added in code so that a dungeon character can only equip an item if they
meet the requirements. Those requirements being: Dextarity and Stamina.
Also the attack of the weapon is added on to the heroes attack and the
defense of his equipment is added into his defense when attacked.