More redundant clean up
I cleaned up the battle the rest of the way such that it has all
printing of characters done using the CharacterPrint.

Additionally I added a print of the order of the battle including who is
going currently and who is next.

Next update I plan to have leveling implemented.